pyproject.toml
[tool.poetry]
name = "crewaicrashcourse"
version = "0.1.0"
description = ""
authors = ["ChakerElj <eljchaker232@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
crewai = "0.1.24"
unstructured = '==0.10.25'
pyowm = '3.3.0'
python-dotenv = "1.0.0"
langchain-openai = "^0.0.5"
[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
